+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Compressed results of `getAllSites` to reduce tool output size.
 - Compressed results of `getAllSiteGroups` to reduce tool output size.
 - Compressed results of `getAllCategories` to reduce tool output size.
+- Compressed results of `getAllCategoryGroups` to reduce tool output size.
 
 ### Fixed
 - Fixed console error.